+++ packages/agent/http/CHANGELOG.md
@@ -7,6 +7,10 @@ and this package adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `@hono/node-server` bumped from `^1.19.12` to `^2.0.1`; `hono` to `^4.12.16`
+
 ## [0.18.6] - 2026-04-10
 
 ### Added